chore(pr-cleanup): restore admin-users files from origin/main into reorganized admin-users-table folder; remove duplicate top-level admin-listings.container.graphql
