--- apps/ui-sharethrift/src/components/layouts/home/account/admin-dashboard/components/admin-listings.container.graphql
@@ -1,32 +1,0 @@
-#import './admin-listings-table.fragment.graphql'
-
-query AdminListingsTableContainerAdminListings(
-  $page: Int!
-  $pageSize: Int!
-  $searchText: String
-  $statusFilters: [String!]
-  $sorter: SorterInput
-) {
-  adminListings(
-    page: $page
-    pageSize: $pageSize
-    searchText: $searchText
-    statusFilters: $statusFilters
-    sorter: $sorter
-  ) {
-    items {
-      ...AdminListingFields
-    }
-    total
-    page
-    pageSize
-  }
-}
-
-mutation AdminListingsTableContainerRemoveListing($id: ObjectID!) {
-  removeListing(id: $id)
-}
-
-mutation AdminListingsTableContainerUnblockListing($id: ObjectID!) {
-  unblockListing(id: $id)
-}
